requirements.txt
paho-mqtt>=1.5.1
argparse>=1.4.0
python-dotenv>=0.19.1